Move train configuration to build-logic
